etl-whether/src/requirements.txt
pandas
pandas-gbq
pyarrow
pytz
google-cloud-bigquery
google-cloud-firestore
google-cloud-logging
google-cloud-secret-manager
google-api-python-client
pytest